+++ sdks/node/src/react-entry.ts
@@ -1,3 +1,4 @@
 export * from './tasks/react.js';
+export * from './tasks/react-components.js';
 export * from './tasks/ui.js';
 export * from './tasks/waitpoint-store.js';